// frontend/src/types/api.d.ts
export interface components {
    schemas: {
        Product: {
            /** @example 1 */
            id?: number;
            /** @example Sticker Pack Vol. 1 (Chibi Style) */
            name?: string;
            /** @example 15000 */
            price?: number;
            /** @example Stiker lucu format PNG & PDF resolusi tinggi. */
            description?: string;
            /**
             * @description URL gambar utama (diambil dari images[0])
             * @example http://localhost:5000/uploads/1700000000-file.png
             */
            image_url?: string;
            /**
             * @example [
             *       "filename1.png",
             *       "filename2.png"
             *     ]
             */
            images?: string[];
            /**
             * @description 1 = Active/Published, 0 = Inactive/Draft
             * @example 1
             */
            is_active?: number;
            /**
             * @description URL link ke halaman produk di Trakteer (opsional)
             * @example https://trakteer.id/inkvoid/shop/product-name-AbCdE
             */
            trakteer_link?: string | null;
        };
        Transaction: {
            /** @example LUMA-173859... */
            order_id?: string;
            /** @example 15000 */
            gross_amount?: number;
            /** @example Budi */
            customer_name?: string;
            /** @example settlement */
            transaction_status?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}